geo-agent/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
aiokafka>=0.8.1
httpx>=0.24.1
asyncpg>=0.28.0
python-dotenv>=1.0.0
